// models/cdd_sdk/generated/cdd_sdktypescript/models/GetConfigurationResponseContent.ts
/**
 * CddService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ConfigurationData } from '../models/ConfigurationData';
import { ErrorDetails } from '../models/ErrorDetails';
import { HttpFile } from '../http/http';

export class GetConfigurationResponseContent {
    'success': boolean;
    'state': string;
    'message': string;
    'error'?: ErrorDetails;
    'configuration'?: ConfigurationData;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "success",
            "baseName": "success",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "state",
            "baseName": "state",
            "type": "string",
            "format": ""
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string",
            "format": ""
        },
        {
            "name": "error",
            "baseName": "error",
            "type": "ErrorDetails",
            "format": ""
        },
        {
            "name": "configuration",
            "baseName": "configuration",
            "type": "ConfigurationData",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetConfigurationResponseContent.attributeTypeMap;
    }

    public constructor() {
    }
}