// models/cdd_sdk/generated/cdd_sdktypescript/models/ReportActualConfigurationRequestContent.ts
/**
 * CddService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DeviceConfiguration } from '../models/DeviceConfiguration';
import { HttpFile } from '../http/http';

export class ReportActualConfigurationRequestContent {
    'configuration': DeviceConfiguration;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "configuration",
            "baseName": "configuration",
            "type": "DeviceConfiguration",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ReportActualConfigurationRequestContent.attributeTypeMap;
    }

    public constructor() {
    }
}